lib-faker/requirements.txt
pytest
pytest-faker